# GH-36059: [C++][Compute] Reserve space for hashtable for scalar lookup functions