--- a/cpp/src/arrow/compute/kernels/scalar_set_lookup_benchmark.cc
+++ b/cpp/src/arrow/compute/kernels/scalar_set_lookup_benchmark.cc
@@ -51,6 +51,7 @@ static void SetLookupBenchmarkString(benchmark::State& state,
   }
   state.SetItemsProcessed(state.iterations() * array_length);
   state.SetBytesProcessed(state.iterations() * values->data()->buffers[2]->size());
+  state.counters["value_set_length"] = value_set_length;
 }
 
 template <typename Type>
@@ -72,6 +73,7 @@ static void SetLookupBenchmarkNumeric(benchmark::State& state,
   }
   state.SetItemsProcessed(state.iterations() * array_length);
   state.SetBytesProcessed(state.iterations() * values->data()->buffers[1]->size());
+  state.counters["value_set_length"] = value_set_length;
 }
 
 static void IndexInStringSmallSet(benchmark::State& state) {
@@ -90,53 +92,56 @@ static void IsInStringLargeSet(benchmark::State& state) {
   SetLookupBenchmarkString(state, "is_in_meta_binary", 1 << 10);
 }
 
+constexpr int64_t kArrayLengthWithSmallSet = 1 << 18;
+constexpr int64_t kArrayLengthWithLargeSet = 1000;
 static void IndexInInt8SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int8Type>(state, "index_in_meta_binary", state.range(0),
-                                      1 << 18);
+                                      kArrayLengthWithSmallSet);
 }
 
 static void IndexInInt16SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int16Type>(state, "index_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IndexInInt32SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int32Type>(state, "index_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IndexInInt64SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int64Type>(state, "index_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IndexInInt32LargeSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int32Type>(state, "index_in_meta_binary", state.range(0),
-                                       1000);
+                                       kArrayLengthWithLargeSet);
 }
 
 static void IsInInt8SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int8Type>(state, "is_in_meta_binary", state.range(0),
-                                      1 << 18);
+                                      kArrayLengthWithSmallSet);
 }
 
 static void IsInInt16SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int16Type>(state, "is_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IsInInt32SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int32Type>(state, "is_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IsInInt64SmallSet(benchmark::State& state) {
   SetLookupBenchmarkNumeric<Int64Type>(state, "is_in_meta_binary", state.range(0),
-                                       1 << 18);
+                                       kArrayLengthWithSmallSet);
 }
 
 static void IsInInt32LargeSet(benchmark::State& state) {
-  SetLookupBenchmarkNumeric<Int32Type>(state, "is_in_meta_binary", state.range(0), 1000);
+  SetLookupBenchmarkNumeric<Int32Type>(state, "is_in_meta_binary", state.range(0),
+                                       kArrayLengthWithLargeSet);
 }
 
 BENCHMARK(IndexInStringSmallSet)->RangeMultiplier(4)->Range(2, 64);
@@ -151,12 +156,12 @@ BENCHMARK(IndexInInt8SmallSet)->RangeMultiplier(4)->Range(2, 8);
 BENCHMARK(IndexInInt16SmallSet)->RangeMultiplier(4)->Range(2, 64);
 BENCHMARK(IndexInInt32SmallSet)->RangeMultiplier(4)->Range(2, 64);
 BENCHMARK(IndexInInt64SmallSet)->RangeMultiplier(4)->Range(2, 64);
-BENCHMARK(IndexInInt32LargeSet)->RangeMultiplier(10)->Range(100, 10000000);
+BENCHMARK(IndexInInt32LargeSet)->RangeMultiplier(100)->Range(100, 1000000);
 BENCHMARK(IsInInt8SmallSet)->RangeMultiplier(4)->Range(2, 8);
 BENCHMARK(IsInInt16SmallSet)->RangeMultiplier(4)->Range(2, 64);
 BENCHMARK(IsInInt32SmallSet)->RangeMultiplier(4)->Range(2, 64);
 BENCHMARK(IsInInt64SmallSet)->RangeMultiplier(4)->Range(2, 64);
-BENCHMARK(IsInInt32LargeSet)->RangeMultiplier(10)->Range(100, 10000000);
+BENCHMARK(IsInInt32LargeSet)->RangeMultiplier(100)->Range(100, 1000000);
 
 }  // namespace compute
 }  // namespace arrow